refactor: split header components in modules
